encryption: Switch from secretbox to aead, authenticate the salt
It might not be entirely necessary (a different salt should cause the derived
key not to match, and the decryption to fail), but I couldn't easily construct
a more-formal argument why there is no malleability issue with the header.

Using the header as associated data fully addresses those concerns, with
essentially no overhead, so I don't see a reason not to do it.